MY-261: добавила текущую задачу и распределение нагрузки для администратора в разделе сформировать задачу
