+++ resources/js/components/FreeRooms.jsx
@@ -1,27 +1,24 @@
-import React, { useEffect, useState } from 'react';
+import React, {useEffect, useRef, useState} from 'react';
 import star from "../../img/star3.svg";
 import 'slick-carousel/slick/slick.css';
 import 'slick-carousel/slick/slick-theme.css';
 import 'react-slideshow-image/dist/styles.css'
 import { Fade } from 'react-slideshow-image';
 import star_full from '../../img/star-full.svg';
+import {useNavigate} from "react-router-dom";
 
 
 export default function FreeRooms(props) {
-    const { checkinDate, checkoutDate, freeRooms, isEditing, callBack } = props;
 
+    const navigate = useNavigate();
 
-
-    console.log(checkinDate, checkoutDate);
-
+    const { checkinDate, checkoutDate, freeRooms, onUpdateFreeRooms} = props;
     const [selectedRoom, setSelectedRoom] = useState(null);
     const [selectedRoomComfort, setSelectedRoomComfort] = useState(null);
     const [isModalOpen, setModalOpen] = useState(false);
     const [isBooking, setIsBooking] = useState(false);
-    const [success, setSuccess] = useState(false);
+    const [response, setResponse] = useState(false);
     const [bookingNumber, setBookingNumber] = useState('');
-    const [showRoom, setShowRoom] = useState(true);
-
 
     const [bookingData, setBookingData] = useState({
         checkinDate: "",
@@ -31,6 +28,32 @@ export default function FreeRooms(props) {
         last_name: "",
         phone: ""
     });
+
+
+        const [isDragging, setIsDragging] = useState(false);
+        const modalRef = useRef(null);
+        const [position, setPosition] = useState({ x: 0, y: 0 });
+
+        const handleMouseDown = (e) => {
+            setIsDragging(true);
+            const offsetX = e.clientX - modalRef.current.getBoundingClientRect().left;
+            const offsetY = e.clientY - modalRef.current.getBoundingClientRect().top;
+            setPosition({ x: offsetX, y: offsetY });
+        };
+
+        const handleMouseMove = (e) => {
+            if (isDragging) {
+                const x = e.clientX - position.x;
+                const y = e.clientY - position.y;
+                modalRef.current.style.left = `${x}px`;
+                modalRef.current.style.top = `${y}px`;
+            }
+        };
+
+        const handleMouseUp = () => {
+            setIsDragging(false);
+        };
+
 
     useEffect(() => {
         setBookingData(prevState => ({
@@ -63,18 +86,17 @@ export default function FreeRooms(props) {
         });
 
         if (response.ok) {
-            // Обработка успешного ответа от сервера
             const data = await response.json();
             setBookingNumber(data);
             console.log(data);
         } else {
-            // Обработка ошибки
             console.log(response.body)
             console.error('Ошибка при выполнении fetch-запроса');
         }
 
         setIsBooking(false);
-        setSuccess(true);
+        setResponse(true);
+
     }
 
 
@@ -91,13 +113,14 @@ export default function FreeRooms(props) {
         setModalOpen(false);
     };
 
-    const checkEditing = (room) => {
-        if (!isEditing)
+    const showBookingForm = () => {
             setIsBooking(true);
-        else {
-            callBack(room);
             setModalOpen(false);
-        }
+    }
+
+    const closeBookingResponse = () => {
+        setResponse(false);
+        onUpdateFreeRooms();
 
     }
 
@@ -185,7 +208,7 @@ export default function FreeRooms(props) {
                             <button
                                 className="btn w-25 align-self-end btn-task rounded-pill px-3"
                                 type="button"
-                                onClick={() => checkEditing(selectedRoom)}
+                                onClick={() => showBookingForm(selectedRoom)}
                             >Забронировать
                             </button>
                         </div>
@@ -194,80 +217,99 @@ export default function FreeRooms(props) {
             </div>
         )}
         {isBooking && (
-            <div className="modal" style={{ display: "block" }}>
-                <div className="modal-dialog">
-                    <div className="modal-content">
-                        <div className="modal-header">
-                            <h5 className="modal-title">Форма бронирования</h5>
-                            <button type="button" className="btn-close" onClick={() => setIsBooking(false)}></button>
-                        </div>
-                        <div className="modal-body">
-                            <div className="form-group">
-                                <label htmlFor="firstName">Имя:</label>
-                                <input
-                                    type="text"
-                                    id="firstName"
-                                    name="first_name"
-                                    onChange={handleBookingDataChange}
-                                    className="form-control"
-                                />
+            <div className="overlay">
+                <div
+                    className="modal"
+                    ref={modalRef}
+                    style={{ display: "block" }}
+                    onMouseDown={handleMouseDown}
+                    onMouseMove={handleMouseMove}
+                    onMouseUp={handleMouseUp}
+                >
+                    <div className="modal-dialog">
+                        <div className="modal-content">
+                            <div className="modal-header">
+                                <h5 className="modal-title">Форма бронирования</h5>
+                                <button type="button" className="btn-close" onClick={() => setIsBooking(false)}></button>
                             </div>
-                            <div className="form-group">
-                                <label htmlFor="lastName">Фамилия:</label>
-                                <input
-                                    type="text"
-                                    id="lastName"
-                                    name="last_name"
-                                    onChange={handleBookingDataChange}
-                                    className="form-control"
-                                />
+                            <div className=" d-flex flex-column modal-body">
+                                <div className="form-group">
+                                    <label htmlFor="firstName">Имя:</label>
+                                    <input
+                                        type="text"
+                                        id="firstName"
+                                        name="first_name"
+                                        onChange={handleBookingDataChange}
+                                        className="form-control"
+                                    />
+                                </div>
+                                <div className="form-group">
+                                    <label htmlFor="lastName">Фамилия:</label>
+                                    <input
+                                        type="text"
+                                        id="lastName"
+                                        name="last_name"
+                                        onChange={handleBookingDataChange}
+                                        className="form-control"
+                                    />
+                                </div>
+                                <div className="form-group">
+                                    <label htmlFor="phone">Телефон:</label>
+                                    <input
+                                        type="tel"
+                                        id="phoneNumber"
+                                        name="phone"
+                                        onChange={handleBookingDataChange}
+                                        className="form-control"
+                                    />
+                                </div>
                             </div>
-                            <div className="form-group">
-                                <label htmlFor="phone">Телефон:</label>
-                                <input
-                                    type="tel"
-                                    id="phoneNumber"
-                                    name="phone"
-                                    onChange={handleBookingDataChange}
-                                    className="form-control"
-                                />
+                            <div className="d-flex p-3 justify-content-end gap-3">
+                                <button
+                                    type="button"
+                                    className="btn btn-border-purple"
+                                    onClick={() => setIsBooking(false)}
+                                >
+                                    Отмена
+                                </button>
+                                <button
+                                    type="button" className="btn btn-task btn-primary"
+                                    onClick={handleBookingSubmit}
+                                >
+                                    Забронировать
+                                </button>
                             </div>
-                        </div>
-                        <div className="modal-footer">
-                            <button
-                                type="button"
-                                className="btn btn-secondary"
-                                onClick={() => setIsBooking(false)}
-                            >
-                                Отмена
-                            </button>
-                            <button
-                                type="button" className="btn btn-primary"
-                                onClick={handleBookingSubmit}
-                            >
-                                Забронировать
-                            </button>
                         </div>
                     </div>
                 </div>
             </div>
+
         )}
 
-        {success && (
-            <div className="modal" style={{ display: "block" }}>
-                <div className="modal-dialog" role="document">
-                    <div className="modal-content rounded-3 shadow">
-                        <div className="modal-body p-4 text-center">
-                            <h3 className="mb-0">Ваша бронь успешно добавлена</h3>
-                            <p className="mb-0">Номер брони: {bookingNumber.bookingId} </p>
+        {response && (
+            <div className="overlay">
+                <div className="modal"
+                     ref={modalRef}
+                     style={{ display: "block" }}
+                     onMouseDown={handleMouseDown}
+                     onMouseMove={handleMouseMove}
+                     onMouseUp={handleMouseUp}
+                >
+                    <div className="modal-dialog" role="document">
+                        <div className="modal-content rounded-3 shadow">
+                            <div className="modal-body p-4 text-center">
+                                <h3 className="mb-0">Ваша бронь успешно добавлена</h3>
+                                <p className="mb-0">Номер брони: {bookingNumber.bookingId} </p>
 
-                        </div>
-                        <div className="modal-footer flex-nowrap p-0">
-                            <button type="button"
-                                className="btn btn-lg btn-link fs-6 text-decoration-none col-6 py-3 m-0 rounded-0 border-end"
-                                onClick={() => setSuccess(false)}
-                            >
-                                <strong>Закрыть</strong></button>
+                            </div>
+                            <div className="modal-footer flex-nowrap p-0">
+                                <button type="button"
+                                        className="btn btn-lg text-purple fs-6 text-decoration-none col-6 py-3 m-0 rounded-0 border-end"
+                                        onClick={closeBookingResponse}
+                                >
+                                    <strong>Закрыть</strong>
+                                </button>
+                            </div>
                         </div>
                     </div>
                 </div>
