+++ resources/js/routes/AddBooking.jsx
@@ -50,12 +50,10 @@ export default function AddBooking() {
             });
 
             if (response.ok) {
-                // Обработка успешного ответа от сервера
                 const data = await response.json();
                 setFreeRooms(data);
                 console.log(data);
             } else {
-                // Обработка ошибки
                 console.error('Ошибка при выполнении fetch-запроса');
             }
         } catch (error) {
@@ -65,6 +63,10 @@ export default function AddBooking() {
 
     const handleShowGuestsAmount = () => {
         setShowGuestsAmount((prevState) => (!prevState));
+    }
+
+    const updateFreeRooms = () => {
+        setFreeRooms([]);
     }
 
     return (
@@ -186,7 +188,7 @@ export default function AddBooking() {
                     </div>
                 </div>
             </div>
-            <FreeRooms checkinDate={checkinDate} checkoutDate={checkoutDate} freeRooms={freeRooms}></FreeRooms>
+            <FreeRooms checkinDate={checkinDate} checkoutDate={checkoutDate} freeRooms={freeRooms} onUpdateFreeRooms={updateFreeRooms}></FreeRooms>
         </div>
 
     );
