+++ resources/sass/app.scss
@@ -51,9 +51,9 @@ html {
 }
 
 .bright-red-button {
-    background-color: #FF4C51;
+    background-color: #F44336;
     color: white;
-    border-color: #FF4C51;
+    border-color: #F44336;
     border-radius: 0 4px 4px 0;
 }
 
@@ -65,9 +65,9 @@ html {
 }
 
 .bright-green-button {
-    background-color: #56CA00;
+    background-color: #4CAF50;
     color: white;
-    border-color: #56CA00;
+    border-color: #4CAF50;
     border-radius: 4px 0 0 4px;
 }
 
@@ -361,12 +361,15 @@ button:hover {
 
 .btn-task {
     background-color: rgb(160, 119, 234);
+    border-radius: 8px;
+    border-color: rgb(160, 119, 234);
     color: white;
 }
 
 .btn-task:hover {
     text-decoration: none;
     background-color: rgb(128, 75, 223);
+    border-color: rgb(160, 119, 234);
     box-shadow: rgba(58, 53, 65, 0.56) 0 6px 18px -8px;
     color: white;
 }
@@ -374,7 +377,8 @@ button:hover {
 .btn-purple-angle {
     background-color: rgb(160, 119, 234);
     color: white;
-    border-radius:  0 8px 8px 0 ;
+    border-radius:  0 8px 8px 0;
+    border-color: rgb(160, 119, 234);
 }
 
 .btn-purple-angle:hover {
@@ -440,6 +444,9 @@ button:hover {
 
 .text-purple {
     color: rgb(128, 75, 223);
+}
+.text-red {
+    color: #c53a3e;
 }
 
 .text-gray {
@@ -605,6 +612,10 @@ button:hover {
     box-shadow: rgba(58, 53, 65, 0.56) 0 6px 18px -8px;
     color: white;
     transform: scale(1.1);
+}
+.task-box {
+    border-right: 1px solid #bfbfbf;
+    padding-right: 50px;
 }
 
 .btn-add:hover .link {
@@ -818,15 +829,13 @@ textarea {
     text-overflow: ellipsis;
 }
 
-//button {
-//    background-color: #4caf50;
-//    color: white;
-//    padding: 10px 15px;
-//    border: none;
-//    border-radius: 4px;
-//    cursor: pointer;
-//}
-//
-//button:hover {
-//    background-color: #45a049;
-//}
+.overlay {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    background-color: rgba(255, 255, 255, 0.5); /* Белый цвет с 50% прозрачности */
+    z-index: 1000; /* Установите z-index так, чтобы ваш фон находился под модальным окном */
+}
+
